Fixing untie to be used only as needed and automatic
when models had shared in/out embedding, had to call untie function to
separate them. This caused increased memory and was applied at all times
regardless of whether those layers were targeted for modification.

change: automatically detect when transform or mixin needs to untie
shared embeddings.

This also adds try except to the untieing code so that if it is invoked
on a model that can't be untied, it gives a warning rather than erroring

new tests are added to test this functionality, old tests are modified
to use the automatic untieing

TODO: offloading tests, ok to delete?

Signed-off-by: HDCharles <charlesdavidhernandez@gmail.com>