Modernize entrypoints module with type hints and use generic types (#1965)
SUMMARY:
This is part of #1927 

Modernize type annotations using | operator and built-in generics in the
`entrypoints module` as part of codebase modernization effort. I also
fixed 2 typos as part of this PR.

TEST PLAN:

- make style
- make quality
- make tests

Notes: 
Happy to address any comments! Also please let me know if you dont want
typos to be fixed as part of this PR. Thank you!