Fix Qwen3Next import errors on old transformers versions (#2030)
SUMMARY:
Currently, we support `transformers>=4.54.0`, however `Qwen3Next` was
only added in `transformers==4.57.0`. This updates the code so that:
1. In `src/llmcompressor/modeling/qwen3_next_moe.py` we only used the
imports for typing, so they've been moved into an `if TYPE_CHECKING`
block.
2. In `tests/llmcompressor/modeling/test_calib_qwen3_next.py` we wrap
the import statement in a try-except and skip the Qwen3Next test if they
fail to import.


TEST PLAN:
Run CI.

---------

Signed-off-by: Fynn Schmitt-Ulms <fschmitt@redhat.com>
Co-authored-by: Dipika Sikka <dipikasikka1@gmail.com>