Use unique filenames for autoround, gptq tests (#2052)
SUMMARY:
Ensure all test files (in `tests/llmcompressor`) have unique filenames
so pytest can be used to run the entire folder at once.


TEST PLAN:
In order to be able to run the entire `tests/llmcompressor` folder of
tests in one go, all test files must have unique filenames. Currently,
that is not the case:

```
$ pytest --co ./tests/llmcompressor/
…
________________________ ERROR collecting tests/llmcompressor/transformers/gptq/test_oneshot.py _________________________
import file mismatch:
imported module 'test_oneshot' has this __file__ attribute:
  /workspaces/llm-compressor/tests/llmcompressor/transformers/autoround/test_oneshot.py
which is not the same as the test file we want to collect:
  /workspaces/llm-compressor/tests/llmcompressor/transformers/gptq/test_oneshot.py
HINT: remove __pycache__ / .pyc files and/or use a unique basename for your test file modules
…
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 2 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
========================================= 539 tests collected, 2 errors in 1.55s =========================================
```

With the changes in this PR:

```
$ pytest --co ./tests/llmcompressor/
…
============================================== 546 tests collected in 1.33s ==============================================
```

Signed-off-by: Domenic Barbuzzi <dbarbuzz@redhat.com>