[AWQ] use match_modules_set and fix logic (#2070)
Depends on https://github.com/vllm-project/compressed-tensors/pull/524

Summary:
- modified AWQ _set_resolved_mappings
- get smoothing and balance layers at same time using match_modules_set
- (bugfix) correct logic so that if any balance layers are incompatible,
that matching is skipped
  -  added warnings
  -  get rid of tqdm and skip counting @kylesayrs 
  -  added helper for module_to_name
- remove hardcoded handling for single balance layer by updating
get_lowest_common_module to handle that
- modified SmoothQuant _resolve_mappings
  - brought into alignment with AWQ
- this is largely a horizontal move though there is handling for
situations that would have been missed before like
      - multiple smooth layer matches in a single set 
      - parent contexts further than 1 layer away.
- updated mapping definitions to always be tuple(list[str],str) which is
always the case but wasn't required unlike in AWQ
- removed get_lowest_common_parent
- now we can use CT's get_lowest_common_ancestor_name so only need to
check for module_list (it has a lot of bugfixes compared to the
get_lowest_common_parent implementation in LLMC)
- updated test_base for AWQ and smoothquant
- added test case for _set_resolved_mappings to check that partially
skipped matches are handled correctly
  - added tests for MoE matching being handled correctly
  - added test cases for get_lowest_non_module_list_ancestor
  - imported Linear and used that instead of torch.nn.Linear
- reverted test_pytorch.py for logarithmic_equalizations and smoothquant
- The test was updated in
https://github.com/vllm-project/llm-compressor/pull/2084 by @rahul-tuli
to ignore some modules but in general because of the way the new logic
works, you need to ignore the whole set.
- if you only ignore one element the matching logic would need to
determine whether there's a full set or not *somehow* which it doesn't
do. In the previous logic, this was possible because it was assumed the
whole set had to be siblings of the smooth_layer, but the new util is
trying to be more flexible and so relaxes this assumption which prevents
the same approach from working. If this is a common need, perhaps we can
add a util that checks for a context parent context of size N or
something.

TEST PLAN:
pytest
/home/HDCharles/repos/llm-compressor/tests/llmcompressor/modifiers/awq/test_base.py
pytest
/home/HDCharles/repos/llm-compressor/tests/llmcompressor/modifiers/smoothquant/test_base.py

---------

Signed-off-by: HDCharles <charlesdavidhernandez@gmail.com>
Signed-off-by: HDCharles <39544797+HDCharles@users.noreply.github.com>
Co-authored-by: Kyle Sayers <kylesayrs@gmail.com>
Co-authored-by: Fynn Schmitt-Ulms <fynnsu@outlook.com>