[awq] simplify compute_layer_means (#2128)
SUMMARY:
Previously I added _orient and _reorient functions to handle this, there
was a TODO to use flatten_for_calibration, this was a little tricky
since
A) its difficult to use to reverse the transformation (see initial
commit)
B) it changes the shape to [1, num_H, num_W, block_H*block_W] and while
we can reshape that directly into the shape needed for the normalization
easily enough, we'd like to record the 4D shape [num_H, num_W, block_H,
block_W] which requires yet more steps to get and record.

Thus it seems simpler to do the unflatten and transpose manually,
exactly how we need it, then record shapes, normalize and lastly use the
recorded shapes to undo it all.

TEST PLAN:
pytest
/home/HDCharles/repos/llm-compressor/src/llmcompressor/modifiers/awq/base.py

---------

Signed-off-by: HDCharles <charlesdavidhernandez@gmail.com>